websocket-server/requirements.txt
python-dotenv
pipecat-ai[cartesia,openai,silero,websocket,deepgram]
weave
